refactor: improve maintainability via dead code cleanup, dedup, and file splits
Dead code removal:
- Remove unused CollectUntil, LoadCheckpointAtTurn, ReconstructSnapshotMessages
- Remove unused testutil harness options (WithModel, WithSystemPrompt, etc.)
- Remove dead boolPtr/extractText/trimRunes duplicates

Duplicate code elimination:
- Unify splitLines -> context.SplitLines (2 copies)
- Unify truncateString/truncate -> truncate.TruncateString (3 copies)
- Unify trimRunes -> truncate.TrimRunes (2 copies)
- Replace extractText with msg.ExtractText() method (2 copies)

File splitting (behavior-preserving):
- Split conv.go (1259L) -> types.go + event_parser.go + event_renderer.go
- Split metrics.go (934L) -> metrics.go + metrics_snapshot.go + metrics_aggregate.go
- Split compact.go (1124L) -> compact.go + compact_summary.go + compact_tools.go

God Object treatment:
- Convert 8 rpcApp closure fields to proper methods in rpc_helpers.go
- Remove initHelpers() and closure struct fields from rpcApp
