+++ pkg/session/manager.go
@@ -3,6 +3,7 @@ package session
 import (
 	"encoding/json"
 	"fmt"
+	agentctx "github.com/tiancaiamao/ai/pkg/context"
 	"os"
 	"path/filepath"
 	"sort"
@@ -520,7 +521,7 @@ func countLegacyMessages(sessPath string) int {
 	}
 
 	count := 0
-	for _, line := range splitLines(data) {
+	for _, line := range agentctx.SplitLines(data) {
 		line = bytesTrimSpace(line)
 		if len(line) == 0 {
 			continue
