+++ pkg/skill/formatter.go
@@ -3,6 +3,8 @@ package skill
 import (
 	"fmt"
 	"strings"
+
+	"github.com/tiancaiamao/ai/pkg/truncate"
 )
 
 const (
@@ -99,7 +101,7 @@ func FormatForPrompt(skills []Skill, stats *SkillStatsFile) string {
 	}
 
 	for _, skill := range selected {
-		description := trimRunes(strings.TrimSpace(skill.Description), maxSkillDescriptionRunes)
+		description := truncate.TrimRunes(strings.TrimSpace(skill.Description), maxSkillDescriptionRunes)
 		lines = append(lines, fmt.Sprintf("- **%s**: %s", skill.Name, description))
 	}
 
@@ -128,15 +130,4 @@ func escapeXML(s string) string {
 	s = strings.ReplaceAll(s, "\"", "&quot;")
 	s = strings.ReplaceAll(s, "'", "&apos;")
 	return s
-}
-
-func trimRunes(s string, limit int) string {
-	if limit <= 0 {
-		return s
-	}
-	runes := []rune(s)
-	if len(runes) <= limit {
-		return s
-	}
-	return string(runes[:limit])
 }
